docstring

+++ meegnet/dataloaders.py
@@ -492,13 +492,13 @@ class ContinuousDataset(EpochedDataset):
 
     Parameters
     ----------
-    window : int, optional
+    window : float, optional
         Window size in seconds. Defaults to 2.
     overlap : float, optional
         Overlap between windows (0-1). Defaults to 0.
     offset : int, optional
         Offset in seconds. Defaults to 10.
-    sfreq : int, optional
+    sfreq : float, optional
         Sampling frequency. Defaults to 500.
     n_subjects : int, optional
         Number of subjects. Defaults to None (all subjects).
